Remove deprecated setting from example config.yml
The `card` setting in the example config.yml was used in the first iteration of the card UI.
The code was later changed to infer the fields that appear in the card, instead of having to configure them (see commit 6abdcdf020e76602cb8ac899d36debbc8e146dcb).
The old card settings, however, were left in the example configuration file, which leads to confusion.